added backend permission check

+++ Gym Application/Gym Application/Controllers/ClassSchedulesController.cs
@@ -16,6 +16,7 @@ using Gym_Application.Models;
 using Business_Layer.DTO;
 
 using System.Net.Mail;
+using Gym_Application.Authentication;
 
 namespace Gym_Application.Controllers
 {
@@ -31,6 +32,7 @@ namespace Gym_Application.Controllers
         // GET: api/ClassSchedules
         [EnableCors(origins: "*", headers: "*", methods: "*")]
         [Route("api/ClassSchedules")]
+        [JwtAuthentication]
         [HttpGet]
         public IEnumerable<ClassSchedule> GetClassSchedule()
         {
@@ -40,6 +42,7 @@ namespace Gym_Application.Controllers
         // GET: api/ClassSchedules/5
         [EnableCors(origins: "*", headers: "*", methods: "*")]
         [ResponseType( typeof( ClassSchedule ) )]
+        [JwtAuthentication]
         public IHttpActionResult GetClassSchedule( int id )
         {
             ClassSchedule classSchedule = service.findOne( id );
@@ -54,9 +57,14 @@ namespace Gym_Application.Controllers
         // POST: api/ClassSchedules/details
         [EnableCors(origins: "*", headers: "*", methods: "*")]
         [Route("api/ClassSchedules/details")]
+        [JwtAuthentication]
         [HttpPost]
         public IHttpActionResult GetClassSchedulesDetails([FromBody] DateSpan dateSpan)
         {
+            // at least trainer
+            if (!Utils.CheckPermission(new List<Role> { Role.ADMIN, Role.TRAINER }))
+                return StatusCode(HttpStatusCode.Forbidden);
+
             try
             {
                 if (ModelState.IsValid)
@@ -76,9 +84,14 @@ namespace Gym_Application.Controllers
         // POST: api/ClassSchedules/trainers/{id}
         [EnableCors(origins: "*", headers: "*", methods: "*")]
         [Route("api/ClassSchedules/trainers/{id}")]
+        [JwtAuthentication]
         [HttpPost]
         public IHttpActionResult GetTrainerScheduleDetails([FromBody] DateSpan dateSpan,int id)
         {
+            // at least trainer
+            if (!Utils.CheckPermission(new List<Role> { Role.ADMIN, Role.TRAINER }))
+                return StatusCode(HttpStatusCode.Forbidden);
+
             try
             {
                 if (ModelState.IsValid)
@@ -100,6 +113,7 @@ namespace Gym_Application.Controllers
         [EnableCors(origins: "*", headers: "*", methods: "*")]
         [ResponseType( typeof( IEnumerable<int> ) )]
         [Route( "api/ClassSchedules/{id}/participants" )]
+        [JwtAuthentication]
         public IHttpActionResult GetClassScheduleParticipants( int id )
         {
             ClassSchedule classSchedule = service.findOne( id );
@@ -116,9 +130,14 @@ namespace Gym_Application.Controllers
         // PUT: api/ClassSchedules/5
         [EnableCors(origins: "*", headers: "*", methods: "*")]
         [ResponseType( typeof( void ) )]
+        [JwtAuthentication]
         public IHttpActionResult PutClassSchedule( int id, ClassSchedule classSchedule )
         {
-            if( !ModelState.IsValid )
+            // at least trainer
+            if (!Utils.CheckPermission(new List<Role> { Role.ADMIN, Role.TRAINER }))
+                return StatusCode(HttpStatusCode.Forbidden);
+
+            if ( !ModelState.IsValid )
             {
                 return BadRequest( ModelState );
             }
@@ -140,9 +159,14 @@ namespace Gym_Application.Controllers
         [ResponseType( typeof( void ) )]
         [HttpPost]
         [Route( "api/ClassSchedules/{id_class_schedule}/participants/{id_user}" )]
+        [JwtAuthentication]
         public IHttpActionResult PostClassScheduleParticipants( int id_class_schedule, int id_user )
         {
-            if( !ModelState.IsValid )
+            // at least user
+            if (!Utils.CheckPermission(new List<Role> { Role.USER, Role.ADMIN, Role.TRAINER }))
+                return StatusCode(HttpStatusCode.Forbidden);
+
+            if ( !ModelState.IsValid )
             {
                 return BadRequest( ModelState );
             }
@@ -168,9 +192,14 @@ namespace Gym_Application.Controllers
         [ResponseType( typeof( void ) )]
         [HttpGet]
         [Route( "api/ClassSchedules/{id_class_schedule}/participants/{id_user}" )]
+        [JwtAuthentication]
         public IHttpActionResult DeleteClassScheduleParticipants( int id_class_schedule, int id_user )
         {
-            if( !ModelState.IsValid )
+            // at least user
+            if (!Utils.CheckPermission(new List<Role> { Role.USER, Role.ADMIN, Role.TRAINER }))
+                return StatusCode(HttpStatusCode.Forbidden);
+
+            if ( !ModelState.IsValid )
             {
                 return BadRequest( ModelState );
             }
@@ -191,10 +220,14 @@ namespace Gym_Application.Controllers
         [EnableCors(origins: "*", headers: "*", methods: "*")]
         [ResponseType( typeof( ClassSchedule ) )]
         [Route("api/ClassSchedules")]
+        [JwtAuthentication]
         public IHttpActionResult PostClassSchedule( ClassSchedule classSchedule )
         {
+            // at least trainer
+            if (!Utils.CheckPermission(new List<Role> { Role.ADMIN, Role.TRAINER }))
+                return StatusCode(HttpStatusCode.Forbidden);
 
-            if( !ModelState.IsValid )
+            if ( !ModelState.IsValid )
             {
                 return BadRequest( ModelState );
             }
@@ -217,8 +250,13 @@ namespace Gym_Application.Controllers
         // DELETE: api/ClassSchedules/5
         [EnableCors(origins: "*", headers: "*", methods: "*")]
         [ResponseType( typeof( void ) )]
+        [JwtAuthentication]
         public IHttpActionResult DeleteClassSchedule( int id )
         {
+            // at least trainer
+            if (!Utils.CheckPermission(new List<Role> { Role.ADMIN, Role.TRAINER }))
+                return StatusCode(HttpStatusCode.Forbidden);
+
             service.delete( id );
             return Ok();
         }
