+++ Gym Application/Gym Application/Controllers/PersonalScheduleController.cs
@@ -13,6 +13,7 @@ using Business_Layer.Services;
 using Gym_Application.Models;
 using Business_Layer.Mappers;
 using System.Web.Http.Results;
+using Gym_Application.Authentication;
 
 namespace Gym_Application.Controllers
 {
@@ -24,6 +25,7 @@ namespace Gym_Application.Controllers
         //merge 
         [EnableCors(origins: "*", headers: "*", methods: "*")]
         [Route("api/PersonalSchedules/")]
+        [JwtAuthentication]
         [HttpGet]
         //[ResponseType(typeof(PersonalSchedule))]
         public IEnumerable<PersonalScheduleView> GetPersonalSchedule()
@@ -40,6 +42,7 @@ namespace Gym_Application.Controllers
         [EnableCors(origins: "*", headers: "*", methods: "*")]
         [ResponseType(typeof(PersonalScheduleView))]
         [Route("api/PersonalSchedules/{id}")]
+        [JwtAuthentication]
         [HttpGet]
         public IHttpActionResult GetPersonalSchedule([FromUri] int id)
         {
@@ -53,10 +56,14 @@ namespace Gym_Application.Controllers
         
         [EnableCors(origins: "*", headers: "*", methods: "*")]
         [Route( "api/PersonalSchedules/{id}/details" )]
+        [JwtAuthentication]
         [ResponseType(typeof(IEnumerable<PersonalScheduleView>))]
         [HttpPost]
         public IHttpActionResult GetPersonalSchedulesDetails([FromBody] DateSpan dateSpan, int id )
         {
+            // at least user
+            if (!Utils.CheckPermission(new List<Role> { Role.USER, Role.ADMIN, Role.TRAINER }))
+                return StatusCode(HttpStatusCode.Forbidden);
             try
             {
                 if (ModelState.IsValid)
@@ -77,10 +84,15 @@ namespace Gym_Application.Controllers
         
         [EnableCors(origins: "*", headers: "*", methods: "*")]
         [Route("api/PersonalSchedules/{id}")]
+        [JwtAuthentication]
         [HttpPut]
         [ResponseType(typeof(void))]
         public IHttpActionResult PutPersonalSchedule(int id, PersonalSchedule personalSchedule)
         {
+            // at least user
+            if (!Utils.CheckPermission(new List<Role> { Role.USER, Role.ADMIN, Role.TRAINER }))
+                return StatusCode(HttpStatusCode.Forbidden);
+
             if (!ModelState.IsValid)
             {
                 return BadRequest(ModelState);
@@ -101,10 +113,14 @@ namespace Gym_Application.Controllers
         //merge
         [EnableCors(origins: "*", headers: "*", methods: "*")]
         [Route("api/PersonalSchedules")]
+        [JwtAuthentication]
         [HttpPost]
         [ResponseType(typeof(PersonalScheduleView))]
         public IHttpActionResult PostPersonalSchedule(PersonalSchedule personalSchedule)
         {
+            // at least user
+            if (!Utils.CheckPermission(new List<Role> { Role.USER, Role.ADMIN, Role.TRAINER }))
+                return StatusCode(HttpStatusCode.Forbidden);
 
             if (!ModelState.IsValid)
             {
@@ -132,9 +148,14 @@ namespace Gym_Application.Controllers
         [EnableCors(origins: "*", headers: "*", methods: "*")]
         [ResponseType(typeof(void))]
         [Route("api/PersonalSchedules/{id}")]
+        [JwtAuthentication]
         [HttpDelete]
         public IHttpActionResult DeletePersonalSchedule(int id)
         {
+            // at least user
+            if (!Utils.CheckPermission(new List<Role> { Role.USER, Role.ADMIN, Role.TRAINER }))
+                return StatusCode(HttpStatusCode.Forbidden);
+
             service.Delete(id);
             return Ok();
         }
